+++ packages/aws-amplify-react/__tests__/Auth/Greetings-test.js
@@ -24,7 +24,7 @@ describe('Greetings', () => {
     describe('normal case', () => {
         test('render correctly with authState signedIn', () => {
             const wrapper = shallow(<Greetings/>);
-            for (var i = 0; i < acceptedStates.length; i += 1){
+            for (let i = 0; i < acceptedStates.length; i += 1){
                 wrapper.setProps({
                     authState: acceptedStates[i],
                     theme: 'theme'
@@ -35,7 +35,7 @@ describe('Greetings', () => {
 
         test('render correctly with hide', () => {
             const wrapper = shallow(<Greetings/>);
-            for (var i = 0; i < acceptedStates.length; i += 1){
+            for (let i = 0; i < acceptedStates.length; i += 1){
                 wrapper.setProps({
                     authState: acceptedStates[i],
                     theme: 'theme',
@@ -59,17 +59,17 @@ describe('Greetings', () => {
                     }
                 },
                 authState: 'signedIn'
-            })  
+            });
 
             expect(wrapper).toMatchSnapshot();
-        })
+        });
     });
 
    
     test('render corrently with other authStates', () => {
         const wrapper = shallow(<Greetings/>);
         
-        for (var i = 0; i < deniedStates.length; i += 1){
+        for (let i = 0; i < deniedStates.length; i += 1){
             wrapper.setProps({
                 authState: deniedStates[i],
                 theme: 'theme'
@@ -77,118 +77,5 @@ describe('Greetings', () => {
 
             expect(wrapper).toMatchSnapshot();
         }
-    });
-
-    describe('signOut test', () => {
-        test('happy case', async () => {
-            const wrapper = shallow(<Greetings/>);
-            const greetings = wrapper.instance();
-
-            const spyon = jest.spyOn(Auth, 'signOut').mockImplementationOnce(() => {
-                return Promise.resolve();
-            });
-
-            await greetings.signOut();
-
-            expect(spyon).toBeCalled();
-            spyon.mockClear();
-        });
-
-        test('error case', async () => {
-            const wrapper = shallow(<Greetings/>);
-            const greetings = wrapper.instance();
-
-            const spyon = jest.spyOn(Auth, 'signOut').mockImplementationOnce(() => {
-                return Promise.reject('error');
-            });
-
-            await greetings.signOut();
-
-            expect(spyon).toBeCalled();
-            spyon.mockClear();
-        });
-    });
-
-    describe('google signOut test', () => {
-        test('happy case', async () => {
-            const mockFn = jest.fn();
-
-            window.gapi = {
-                auth2: {
-                    getAuthInstance() {
-                        return Promise.resolve({
-                            signOut: mockFn
-                        })
-                    }
-                }
-            };
-
-            const wrapper = shallow(<Greetings/>);
-            const greetings = wrapper.instance();
-
-            await greetings.googleSignOut();
-
-            expect(mockFn).toBeCalled();
-        });
-
-        test('no auth2', async () => {
-            window.gapi = null;
-            const wrapper = shallow(<Greetings/>);
-            const greetings = wrapper.instance();
-
-            expect(await greetings.googleSignOut()).toBeNull();
-        });
-
-        test('no googleAuth', async () => {
-            window.gapi = {
-                auth2: {
-                    getAuthInstance() {
-                        return Promise.resolve(null);
-                    }
-                }
-            };
-
-            const wrapper = shallow(<Greetings/>);
-            const greetings = wrapper.instance();
-
-            await greetings.googleSignOut();
-        });
-    });
-
-    describe('facebook signout test', () => {
-        test('happy case', async () => {
-            window.FB = {
-                getLoginStatus(callback) {
-                    callback({
-                        status: 'connected'
-                    })
-                },
-                logout(callback) {
-                    callback('response');
-                }
-            }
-
-            const wrapper = shallow(<Greetings/>);
-            const greetings = wrapper.instance();
-            
-            await greetings.facebookSignOut()
-        });
-
-        test('not connected', async () => {
-            window.FB = {
-                getLoginStatus(callback) {
-                    callback({
-                        status: 'not connected'
-                    })
-                },
-                logout(callback) {
-                    callback('response');
-                }
-            }
-            const wrapper = shallow(<Greetings/>);
-            const greetings = wrapper.instance();
-            
-            await greetings.facebookSignOut()
-        });
-    });
+    });   
 });