+++ packages/aws-amplify-react/src/Auth/SignIn.jsx
@@ -56,10 +56,11 @@ export default class SignIn extends AuthPiece {
     }
 
     onKeyDown(e) {
-        if (this.props.authState === 'signIn' && !this.props.hide) {
-            if (e.keyCode === 13) { // when press enter
-                this.signIn();
-            }
+        if (e.keyCode !== 13) return;
+
+        const { hide = [] } = this.props;
+        if (this.props.authState === 'signIn' && !hide.includes(SignIn)) {
+            this.signIn();
         }
     }
 
@@ -103,11 +104,11 @@ export default class SignIn extends AuthPiece {
             .catch(err => {
                 if (err.code === 'UserNotConfirmedException') {
                     logger.debug('the user is not confirmed');
-                    this.changeState('confirmSignUp');
-                } 
+                    this.changeState('confirmSignUp', { username });
+                }
                 else if (err.code === 'PasswordResetRequiredException') {
                     logger.debug('the user requires a new password');
-                    this.changeState('requireNewPassword');
+                    this.changeState('forgotPassword', { username });
                 } else {
                     this.error(err);
                 }
@@ -115,10 +116,10 @@ export default class SignIn extends AuthPiece {
     }
 
     showComponent(theme) {
-        const { authState, hide = [], federated, onStateChange } = this.props;
+        const { authState, hide = [], federated, onStateChange, onAuthEvent, hideLink=[] } = this.props;
         if (hide && hide.includes(SignIn)) { return null; }
-        const hideSignUp = hide.some(component => component.name === 'SignUp')
-        const hideForgotPassword = hide.some(component => component.name === 'ForgotPassword')
+        const hideSignUp = hideLink.some(component => component.name === 'SignUp');
+        const hideForgotPassword =hideLink.some(component => component.name === 'ForgotPassword');
         return (
             <FormSection theme={theme}>
                 <SectionHeader theme={theme}>{I18n.get('Sign in to your account')}</SectionHeader>
@@ -128,6 +129,7 @@ export default class SignIn extends AuthPiece {
                         theme={theme}
                         authState={authState}
                         onStateChange={onStateChange}
+                        onAuthEvent={onAuthEvent}
                     />
                     <FormField theme={theme}>
                         <InputLabel>{I18n.get('Username')} *</InputLabel>
@@ -159,7 +161,7 @@ export default class SignIn extends AuthPiece {
                             </Hint>
                         }
                     </FormField>
-                    
+
                 </SectionBody>
                 <SectionFooter theme={theme}>
                     <SectionFooterPrimaryContent theme={theme}>
@@ -177,6 +179,6 @@ export default class SignIn extends AuthPiece {
                     }
                 </SectionFooter>
             </FormSection>
-        )
+        );
     }
 }